Serve files from process.cwd()
This allows to render markdown files that include relative links to
images or other files hosted on Github.

The patch also replaces the file serving code that was used with the
'send' module, which is the static file serving module that is used by
connect in order to simplify the code.